+++ bindings/pyfiction/include/pyfiction/algorithms/simulation/sidb/operational_domain.hpp
@@ -133,9 +133,9 @@ inline void operational_domain(pybind11::module& m)
         .def_readonly("num_non_operational_parameter_combinations",
                       &fiction::operational_domain_stats::num_non_operational_parameter_combinations,
                       DOC(fiction_operational_domain_stats_num_non_operational_parameter_combinations))
-        .def_readonly("total_number_of_parameter_points",
-                      &fiction::operational_domain_stats::total_number_of_parameter_points,
-                      DOC(fiction_operational_domain_stats_total_number_of_parameter_points))
+        .def_readonly("num_total_parameter_points",
+                      &fiction::operational_domain_stats::num_total_parameter_points,
+                      DOC(fiction_operational_domain_stats_num_total_parameter_points))
 
         ;
