+++ docs/algorithms/sidb_simulation.rst
@@ -252,11 +252,11 @@ Operational Domain Computation
         .. doxygenfunction:: fiction::operational_domain_flood_fill
         .. doxygenfunction:: fiction::operational_domain_contour_tracing
 
-        **Header:** ``fiction/algorithms/simulation/sidb/compute_operational_ratio_at_point.hpp``
+        **Header:** ``fiction/algorithms/simulation/sidb/compute_operational_ratio.hpp``
 
-        .. doxygenstruct:: fiction::compute_operational_ratio_at_point_params
+        .. doxygenstruct:: fiction::compute_operational_ratio
            :members:
-        .. doxygenfunction:: fiction::compute_operational_ratio_at_point
+        .. doxygenfunction:: fiction::compute_operational_ratio
 
     .. tab:: Python
         .. autoclass:: mnt.pyfiction.operational_status
@@ -284,10 +284,10 @@ Operational Domain Computation
         .. autofunction:: mnt.pyfiction.operational_domain_flood_fill
         .. autofunction:: mnt.pyfiction.operational_domain_contour_tracing
 
-        .. autoclass:: mnt.pyfiction.compute_operational_ratio_at_point_params
+        .. autoclass:: mnt.pyfiction.compute_operational_ratio
                 :members:
 
-        .. autofunction:: mnt.pyfiction.compute_operational_ratio_at_point
+        .. autofunction:: mnt.pyfiction.compute_operational_ratio
 
 
 Determine Physically Valid Parameters
